Configure memory, 32-bit WAL.
