+++ vfs/os_std_atomic.go
@@ -1,4 +1,4 @@
-//go:build !linux || !(amd64 || arm64 || riscv64 || ppc64le) || sqlite3_nosys
+//go:build !linux || !(amd64 || arm64 || riscv64) || sqlite3_nosys
 
 package vfs
