+++ vfs/tests/mptest/mptest_test.go
@@ -43,10 +43,9 @@ func TestMain(m *testing.M) {
 	ctx := context.Background()
 	cfg := wazero.NewRuntimeConfig().
 		WithCoreFeatures(api.CoreFeaturesV2 | experimental.CoreFeaturesThreads).
-		WithMemoryLimitPages(1024)
+		WithMemoryLimitPages(512)
 	rt = wazero.NewRuntimeWithConfig(ctx, cfg)
 	wasi_snapshot_preview1.MustInstantiate(ctx, rt)
-
 	env := vfs.ExportHostFunctions(rt.NewHostModuleBuilder("env"))
 	env.NewFunctionBuilder().WithFunc(system).Export("system")
 	_, err := env.Instantiate(ctx)
